One search box, on every page
There were two: HeaderListSearch, which filtered the list under it, and
HeaderSearch, which navigated to the feed on every other page. They shared the
debounce, the stale-response token, the arrow keys, the Cmd+K and `/` hotkeys,
the click-away dismissal and the row rendering — in two copies — and differed in
exactly one thing: what a pick DOES.

That one thing is a target now. Where a list has registered itself the target is
its filter store; everywhere else it is a target that turns the same pick into a
link to the feed carrying the same filter. Nothing below that line asks which
kind of page it is on.

So the suggestions, the composed picks and the empty box's starting points —
which shipped last week on the feed alone — are now on the job page, /about,
/roles, the blog, the profile, everywhere. On a listless page the empty box
fetches the category distribution itself, once, on first focus: a page nobody
searches from should not pay for it.

browseQuery builds the feed URL through the filter model rather than by
concatenating params, so the link a navigation opens and the URL the in-place
filter writes come from one serializer. Two would drift, and the drift would be
a link that filters differently from the control that made it.

TopBar keeps `listKind` only to word the placeholder. It used to select between
the two components.

Both design-system ratchets move because a file was deleted, which is the point:
token violations drop from 509 to 507, and EntityLogo loses the consumer that was
a copy of another one.
