+++ design-system/scripts/adoption-baseline.json
@@ -10,7 +10,7 @@
   "CountryFlag": 10,
   "Dialog": 8,
   "EmptyState": 2,
-  "EntityLogo": 24,
+  "EntityLogo": 23,
   "FormField": 2,
   "Input": 15,
   "LoadMore": 7,
