+++ design-system/scripts/web-token-baseline.json
@@ -41,10 +41,9 @@
   "web/src/lib/components/ForCompaniesView.svelte": 4,
   "web/src/lib/components/GhostChecklist.svelte": 6,
   "web/src/lib/components/GmailConnectDialog.svelte": 1,
-  "web/src/lib/components/HeaderListSearch.svelte": 2,
   "web/src/lib/components/HeaderLocationFilter.svelte": 2,
   "web/src/lib/components/HeaderMenu.svelte": 1,
-  "web/src/lib/components/HeaderSearch.svelte": 1,
+  "web/src/lib/components/HeaderSearch.svelte": 2,
   "web/src/lib/components/HomeFunnel.svelte": 1,
   "web/src/lib/components/HomeView.svelte": 16,
   "web/src/lib/components/InboxLandingView.svelte": 4,
